fix hazard pointers (bad sorting), clean up mpmc_fifo (and make it always safe)

+++ include/lockfree_ring_buffer.h
@@ -86,7 +86,7 @@ static inline size_t lockfree_ring_buffer_size(const lockfree_ring_buffer_t* rb)
 {
     const uint64_t high = rb->high;
     load_load_barrier();//read high first; make it look less than or equal to its actual size
-    const ssize_t size = high - rb->low;
+    const int64_t size = high - rb->low;
     return size >= 0 ? size : 0;
 }
